deployer: code refactoring
