+++ cmd/deploy/main.go
@@ -12,14 +12,20 @@ import (
 
 func main() {
 	cfg := deployer.NewConfig()
+
+	deployer.SetupLogging(cfg)
+
 	grpcConnection, err := deployer.NewGrpcConnection(cfg)
 	if err != nil {
 		log.Errorf("fatal: %s", err)
 		os.Exit(int(deployer.ExitUnavailable))
 	}
 	defer grpcConnection.Close()
+
 	grpcClient := pb.NewDeployClient(grpcConnection)
-	d := deployer.Deployer{Client: grpcClient, DeployServer: cfg.DeployServerURL}
+
+	d := deployer.Deployer{Client: grpcClient}
+
 	code, err := d.Run(cfg)
 
 	if err != nil {
